[chore] 코드래빗 수정반영
