+++ app/editor/[id]/components/EditorUpdate.tsx
@@ -4,6 +4,7 @@ import { notFound } from 'next/navigation';
 import { useEffect } from 'react';
 
 import LoadingSpinner from '@/shared/assets/icons/loadingSpinner.svg';
+import { usePreventBack } from '@/shared/hooks/usePreventBack';
 import LeftPanel from '@/widgets/editor/leftPanel/LeftPanel';
 import Preview from '@/widgets/editor/preview/Preview';
 import RightPanel from '@/widgets/editor/rightPanel/RightPanel';
@@ -32,37 +33,11 @@ function EditorUpdate({ folderId, uuid }: Props) {
       initBgmStore();
     }
   }, [savedData, initEditStore, initBgmStore, initBulkData]);
-  useEffect(() => {
-    // 1. 탭 닫기 / 새로고침 방지
-    const handleBeforeUnload = (e: BeforeUnloadEvent) => {
-      e.preventDefault();
-    };
-
-    // 2. 뒤로가기 방지
-    const handlePopState = () => {
-      const leave = window.confirm(
-        '수정된 내용이 저장되지 않을 수 있습니다.\n정말 나가시겠습니까?'
-      );
-      if (leave) {
-        history.back();
-      } else {
-        history.pushState(null, '', window.location.href);
-      }
-    };
-
-    history.pushState(null, '', window.location.href);
-    window.addEventListener('beforeunload', handleBeforeUnload);
-    window.addEventListener('popstate', handlePopState);
-
-    return () => {
-      window.removeEventListener('beforeunload', handleBeforeUnload);
-      window.removeEventListener('popstate', handlePopState);
-    };
-  }, []);
+  usePreventBack();
   if (error) notFound();
   if (loading || !savedData) {
     return (
-      <div className="w-screen h-screen flex justify-center items-center bg-[#E7E9EB]">
+      <div className="w-screen h-full flex justify-center items-center bg-[#E7E9EB]">
         <LoadingSpinner className="w-20 h-20 animate-spin" />
       </div>
     );
