+++ features/session/components/HomeButton.tsx
@@ -5,10 +5,10 @@ import { usePathname } from 'next/navigation';
 import React from 'react';
 
 function HomeButton() {
-  const pathname = usePathname(); // "/products/123"
+  const pathname = usePathname();
 
   const handleHomeClick = (e: React.MouseEvent) => {
-    if (pathname === '/editor') {
+    if (pathname.startsWith('/editor')) {
       const leave = window.confirm(
         '수정된 내용이 저장되지 않을 수 있습니다.\n정말 나가시겠습니까?'
       );
