+++ widgets/editor/preview/Preview.tsx
@@ -63,7 +63,10 @@ function Preview() {
       id="preview-container"
       className="w-93.75 h-full max-h-[810px] flex flex-col gap-4 relative"
     >
-      <div className="max-h-[750px] bg-white" style={{ backgroundColor }}>
+      <div
+        className="max-h-[750px] h-full bg-white"
+        style={{ backgroundColor }}
+      >
         <div className="overflow-y-auto h-full w-93.75 box-border scrollbar-hide">
           <div className="flex flex-col min-h-full font-maruburi">
             <MainPosterPreview />
