+++ widgets/editor/rightPanel/RightPanel.tsx
@@ -30,7 +30,7 @@ function RightPanel() {
 
   return (
     <div
-      className={`${selectedId !== 'mainPoster' && (!typeArray || typeArray.length === 0) ? 'invisible' : 'visible'} 'w-93.75 h-full max-h-[810px] mr-15 flex flex-col gap-5'`}
+      className={`${selectedId !== 'mainPoster' && (!typeArray || typeArray.length === 0) ? 'invisible' : 'visible'} w-93.75 h-full max-h-[810px] mr-15 flex flex-col gap-5`}
     >
       <div className="w-93.75 min-h-0 flex-1 bg-white rounded-lg shadow-edit flex-center flex-col gap-3 px-5 ">
         <div className="w-full">
